DUOS-926 Remove DataAccessRequestAPI
