+++ src/test/java/org/broadinstitute/consent/http/resources/ElectionReviewResourceTest.java
@@ -8,10 +8,8 @@ import org.broadinstitute.consent.http.models.DataAccessRequestData;
 import org.broadinstitute.consent.http.models.Election;
 import org.broadinstitute.consent.http.models.ElectionReview;
 import org.broadinstitute.consent.http.models.Vote;
-import org.broadinstitute.consent.http.service.AbstractDataAccessRequestAPI;
 import org.broadinstitute.consent.http.service.AbstractReviewResultsAPI;
 import org.broadinstitute.consent.http.service.ConsentService;
-import org.broadinstitute.consent.http.service.DataAccessRequestAPI;
 import org.broadinstitute.consent.http.service.DataAccessRequestService;
 import org.broadinstitute.consent.http.service.ElectionService;
 import org.broadinstitute.consent.http.service.ReviewResultsAPI;
@@ -36,15 +34,12 @@ import static org.mockito.Mockito.when;
 @RunWith(PowerMockRunner.class)
 @PowerMockIgnore("jdk.internal.reflect.*")
 @PrepareForTest({
-        AbstractDataAccessRequestAPI.class,
         AbstractReviewResultsAPI.class
 })
 public class ElectionReviewResourceTest {
 
     @Mock
     private ConsentService consentService;
-    @Mock
-    private DataAccessRequestAPI accessRequestAPI;
     @Mock
     private ElectionService electionService;
     @Mock
@@ -57,12 +52,10 @@ public class ElectionReviewResourceTest {
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
-        PowerMockito.mockStatic(AbstractDataAccessRequestAPI.class);
         PowerMockito.mockStatic(AbstractReviewResultsAPI.class);
     }
 
     private void initResource() {
-        when(AbstractDataAccessRequestAPI.getInstance()).thenReturn(accessRequestAPI);
         when(AbstractReviewResultsAPI.getInstance()).thenReturn(reviewResultsAPI);
         resource = new ElectionReviewResource(darService, consentService, electionService);
     }
